cpu/esp_common: platform independent thread_arch implementation
Xtensa-specific code was moved from thread_arch.c to thread_arch_xtensa.c. RISC-V code has been added to thread_arch_riscv.c. This allows to handle both Xtensa and RISC-V based ESP SoCs.
